+++ .github/workflows/test_and_lint.yml
@@ -27,7 +27,7 @@ jobs:
         poetry install
 
     - name: Run tests
-      run: poetry run pytest tests --cov=minivnn --cov-report=xml --cov-report=term
+      run: poetry run pytest tests --cov=minivn --cov-report=xml --cov-report=term
 
     - name: Check code formatting with Black
       run: poetry run black --check --config pyproject.toml .
