+++ README.md
@@ -1,25 +1,25 @@
-# minivnn
+# minivn
 
-![Tests](https://github.com/aismlv/minivnn/actions/workflows/test_and_lint.yml/badge.svg)
-[![codecov](https://codecov.io/gh/aismlv/minivnn/branch/main/graph/badge.svg?token=5J503UR8O7)](https://codecov.io/gh/aismlv/minivnn)
-[![PyPI version](https://badge.fury.io/py/minivnn.svg)](https://pypi.org/project/minivnn/)
+![Tests](https://github.com/aismlv/minivn/actions/workflows/test_and_lint.yml/badge.svg)
+[![codecov](https://codecov.io/gh/aismlv/minivn/branch/main/graph/badge.svg?token=5J503UR8O7)](https://codecov.io/gh/aismlv/minivn)
+[![PyPI version](https://badge.fury.io/py/minivn.svg)](https://pypi.org/project/minivn/)
 [![License: Apache 2.0](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
 
-`minivnn` (pronounced "minivan" 🚐) is an exact nearest neighbor search Python library for those times when "approximate" just won't cut it (or is simply overkill).
+`minivn` (pronounced "minivan" 🚐) is an exact nearest neighbor search Python library for those times when "approximate" just won't cut it (or is simply overkill).
 
 ## Installation
 
-Install `minivnn` using `pip`:
+Install `minivn` using `pip`:
 
 ```bash
-pip install minivnn
+pip install minivn
 ```
 
 ## Usage
-Here's an example of how to use minivnn:
+Here's an example of how to use minivn:
 
 ```python
-from minivnn import Index
+from minivn import Index
 import numpy as np
 
 # Create an index with 128-dimensional embeddings and cosine similarity metric
