+++ minivn/__init__.py
@@ -1,0 +1,3 @@
+from .minivn import Index
+
+__all__ = ["Index"]
