--- minivnn/__init__.py
@@ -1,3 +1,0 @@
-from .minivnn import Index
-
-__all__ = ["Index"]
